Fix cross-platform compile errors

+++ ProjectEuler0021/ProjectEuler0021.cpp
@@ -12,15 +12,15 @@
 #include "prime_helper.h"
 
 
-std::vector<int32_t> get_amicable_numbers(int max_val) {
+std::vector<int32_t> get_amicable_numbers(uint64_t max_val) {
     PrimeHelper helper;
-    int internal_max = max_val * 2.6;
+    uint64_t internal_max = max_val * 2.6;
 
     std::vector<uint64_t> sums;
     sums.push_back(-1);
     sums.push_back(0);
 
-    for (int i = 2; i < internal_max; ++i) {
+    for (uint64_t i = 2; i < internal_max; ++i) {
         uint64_t sum{ 0 };
         for (const auto& fact : helper.get_proper_divisors(i)) {
             sum += fact;
@@ -29,7 +29,7 @@ std::vector<int32_t> get_amicable_numbers(int max_val) {
     }
 
     std::vector<int32_t> ret;
-    for (int i = 2; i < max_val; ++i) {
+    for (uint64_t i = 2; i < max_val; ++i) {
         uint64_t index1 = sums[i];
         if (index1 == i) {
             std::cout << "Found perfect number: " << i << std::endl;
